Refactor computation engine into facade classes (#547)
- Extract ForecastHistoryStore for forecast persistence
- Extract ForecastPipeline for load/solar/excess-solar computations
- Extract OptimizerFacade for DP optimizer orchestration
- Extract PriceSignalEngine for cheap price and spike logic
- Simplify ComputationEngine.compute_derived_values to orchestration
- Add unit tests for each extracted class
- Remove dead code (legacy spike analysis tests moved to price_signal_engine)

Fixes #547
