+++ custom_components/localshift/computation_engine.py
@@ -15,11 +15,7 @@ from .computation_engine_lib import (
     ExcessSolarSignalsEngine,
     ForecastAccuracyEngine,
     HistoryFetcher,
-    PriceCalculator,
-    SpikeAnalyzer,
     WeatherDiagnosticsEngine,
-    analyze_spike_window,
-    calculate_spike_price_threshold,
     max_forecast_price,
     parse_forecast_dt,
     percentile,
@@ -27,24 +23,16 @@ from .computation_engine_lib import (
     sum_solar_before_target,
 )
 from .computation_engine_lib.excess_solar import ExcessSolarEngine
+from .computation_engine_lib.forecast_history_store import ForecastHistoryStore
+from .computation_engine_lib.forecast_pipeline import ForecastPipeline
 from .computation_engine_lib.load_forecaster import LoadForecaster
 from .computation_engine_lib.optimizer_dp import DPPlanner
-from .computation_engine_lib.optimizer_runner import (
-    OptimizerSafetyGate,
-    _build_optimizer_config,
-    _build_summary,
-    _derive_runtime_apply_plan,
-    _find_current_slot_index,
-    _normalize_initial_soc,
-    _serialize_decision,
-    _serialize_result,
-)
-from .computation_engine_lib.slot_builder import SlotBuilder
+from .computation_engine_lib.optimizer_facade import OptimizerFacade
+from .computation_engine_lib.optimizer_runner import _find_current_slot_index
+from .computation_engine_lib.price_signal_engine import PriceSignalEngine
 from .computation_engine_lib.slot_schedule import TOTAL_SLOTS
 from .computation_engine_lib.soc_simulator import SocSimulator
 from .const import (
-    BATTERY_CAPACITY_KWH,
-    CHARGE_RATE_GRID_KW,
     CONF_ALLOW_DW_ENTRY_UNDER_TARGET,
     CONF_BATTERY_TARGET,
     CONF_DEMAND_WINDOW_END,
@@ -103,8 +91,7 @@ class ComputationEngine:
         self._get_switch_state = get_switch_state_func
 
         # Forecast history storage (HA Storage) - persists predictions across restarts
-        self._forecast_history_store: Any = None  # Initialized in async_start
-        self._forecast_history_loaded: bool = False
+        self._forecast_history_store = ForecastHistoryStore(hass)
 
         # History fetcher for historical load data (delegated to separate module)
         self._history_fetcher = HistoryFetcher(hass, entry)
@@ -125,28 +112,19 @@ class ComputationEngine:
             simulate_with_additional_load=self._soc_simulator._simulate_with_additional_load,
         )
 
-        self._price_calculator = PriceCalculator(
-            entry=entry,
-            parse_forecast_dt=self._parse_forecast_dt,
-            percentile_func=self._percentile,
-            sum_solar_before_target=self._sum_solar_before_target,
-            get_expected_load_kw=self._get_expected_load_kw_from_slots,
-        )
-        self._spike_analyzer = SpikeAnalyzer(
+        self._price_signals = PriceSignalEngine(
             entry=entry,
             get_switch_state=self._get_switch_state,
             parse_time_option=self._parse_time_option,
-            analyze_spike_window=analyze_spike_window,
-            calculate_spike_price_threshold=calculate_spike_price_threshold,
         )
         self._excess_solar_signals = ExcessSolarSignalsEngine(
             entry=entry,
-            calculate_excess_by_windows=self._excess_solar_engine._calculate_excess_by_windows,
-            find_nearest_negative_fit_window=self._excess_solar_engine._find_nearest_negative_fit_window,
-            calculate_excess_until_negative_fit=self._excess_solar_engine._calculate_excess_until_negative_fit,
-            find_battery_fill_point=self._soc_simulator._find_battery_fill_point,
-            calculate_safe_additional_load=self._excess_solar_engine._calculate_safe_additional_load,
-            compute_load_shift_signal=self._excess_solar_engine._compute_load_shift_signal,
+            calculate_excess_by_windows=self._excess_solar_engine.calculate_excess_by_windows,
+            find_nearest_negative_fit_window=self._excess_solar_engine.find_nearest_negative_fit_window,
+            calculate_excess_until_negative_fit=self._excess_solar_engine.calculate_excess_until_negative_fit,
+            find_battery_fill_point=self._soc_simulator.find_battery_fill_point,
+            calculate_safe_additional_load=self._excess_solar_engine.calculate_safe_additional_load,
+            compute_load_shift_signal=self._excess_solar_engine.compute_load_shift_signal,
             get_entity_id=self._get_entity_id,
             get_historical_hourly_averages=self._get_historical_hourly_averages,
             recent_load_1hr_getter=lambda: self._recent_load_1hr_kw,
@@ -155,15 +133,21 @@ class ComputationEngine:
         self._forecast_accuracy = ForecastAccuracyEngine()
         self._weather_diagnostics = WeatherDiagnosticsEngine(entry)
 
+        self._forecast_pipeline = ForecastPipeline(
+            load_forecaster=self._load_forecaster,
+            price_signals=self._price_signals,
+            forecast_history_store=self._forecast_history_store,
+            get_switch_state=self._get_switch_state,
+            excess_solar_signals=self._excess_solar_signals,
+        )
+
         # DP Planner for Phase 3 (#441) - eliminates one-cycle lag
         self._dp_planner = DPPlanner()
 
-        # Solar accuracy tracker for Issue #378/#513 - bias correction
-        self._solar_accuracy_tracker: Any = None
+        self._optimizer_facade = OptimizerFacade(self._dp_planner)
 
         # Local cache properties (delegated to history_fetcher for storage)
         self._previous_active_mode = None
-        self._last_forecast_hour: int | None = None
         self._last_decision_log_time: datetime | None = None
 
         # Baseline load profile for Issue #137 (set by coordinator)
@@ -175,100 +159,8 @@ class ComputationEngine:
         Args:
             tracker: SolarAccuracyTracker instance
         """
-        self._solar_accuracy_tracker = tracker
+        self._optimizer_facade.set_solar_accuracy_tracker(tracker)
         _LOGGER.info("Solar accuracy tracker connected to computation engine")
-
-    def _record_forecasts_for_slots(
-        self,
-        slots: list[Any],
-        weather_condition: str,
-    ) -> None:
-        """Record solar forecasts for all slots with solar production.
-
-        Called after building slots to track forecasts for later bias analysis.
-
-        Args:
-            slots: List of SlotContext objects from slot_builder
-            weather_condition: Current weather condition string
-        """
-        if self._solar_accuracy_tracker is None:
-            return
-
-        recorded = 0
-        for slot in slots:
-            if slot.solar_kwh > 0.01:
-                period_start = datetime.fromisoformat(slot.timestamp_iso)
-                self._solar_accuracy_tracker.record_forecast(
-                    period_start=period_start,
-                    forecast_kwh=slot.solar_kwh,
-                    weather_condition=weather_condition,
-                )
-                recorded += 1
-
-        if recorded > 0:
-            _LOGGER.debug("Recorded %d solar forecasts for accuracy tracking", recorded)
-
-    def _apply_bias_correction_to_slots(
-        self,
-        slots: list[Any],
-        weather_condition: str,
-    ) -> None:
-        """Apply bias correction to solar forecasts for each slot.
-
-        Uses the solar accuracy tracker to get context-specific bias
-        and adjusts solar_kwh accordingly. Rounds down to avoid over-charging.
-
-        Args:
-            slots: List of SlotContext objects from slot_builder
-            weather_condition: Current weather condition string
-        """
-        if self._solar_accuracy_tracker is None:
-            return
-
-        from homeassistant.util import dt as dt_util
-
-        now = dt_util.now()
-        hour = now.hour
-        if 6 <= hour < 12:
-            time_of_day = "morning"
-        elif 12 <= hour < 18:
-            time_of_day = "afternoon"
-        elif 18 <= hour < 21:
-            time_of_day = "evening"
-        else:
-            time_of_day = "night"
-
-        month = now.month
-        if month in (12, 1, 2):
-            season = "summer"
-        elif month in (3, 4, 5):
-            season = "autumn"
-        elif month in (6, 7, 8):
-            season = "winter"
-        else:
-            season = "spring"
-
-        bias_multiplier = self._solar_accuracy_tracker.get_bias_correction(
-            time_of_day, weather_condition, season
-        )
-
-        if bias_multiplier != 1.0:
-            corrected = 0
-            for slot in slots:
-                if slot.solar_kwh > 0.01:
-                    original = slot.solar_kwh
-                    slot.solar_kwh = max(0.0, slot.solar_kwh * bias_multiplier)
-                    if abs(slot.solar_kwh - original) > 0.001:
-                        corrected += 1
-            if corrected > 0:
-                _LOGGER.info(
-                    "Applied bias correction: multiplier=%.2f (%s/%s/%s), corrected %d slots",
-                    bias_multiplier,
-                    time_of_day,
-                    weather_condition,
-                    season,
-                    corrected,
-                )
 
     # ========================================================================
     # MAIN ENTRY POINT
@@ -332,8 +224,12 @@ class ComputationEngine:
         # Compute effective_cheap_price BEFORE forecast using preliminary solar estimate
         # This breaks the circular dependency where forecast depends on effective_cheap_price
         # which depends on solar_can_reach_target which depends on forecast
-        self._compute_effective_cheap_price_preliminary(
-            data, now_dt, before_dw, target_hour, target_pct
+        self._price_signals.compute_effective_cheap_price_preliminary(
+            data=data,
+            now_dt=now_dt,
+            before_dw=before_dw,
+            target_hour=target_hour,
+            target_pct=target_pct,
         )
 
         # Set allow_dw_entry_under_target flag on data for forecast_computer
@@ -349,7 +245,13 @@ class ComputationEngine:
         load_entity_id = self._get_entity_id("teslemetry_load_power")
         hourly_avg_kw = self._get_historical_hourly_averages(load_entity_id)
         recent_load_kw = self._recent_load_1hr_kw
-        self._compute_load_forecast_slots(data, now_dt, hourly_avg_kw, recent_load_kw)
+        self._forecast_pipeline.compute_load_forecast_slots(
+            data=data,
+            now_dt=now_dt,
+            historical_avg_kw=hourly_avg_kw,
+            recent_load_kw=recent_load_kw,
+            total_slots=TOTAL_SLOTS,
+        )
 
         # ---- Bridge HistoryFetcher results to CoordinatorData (Issue #493) ----
         # These fields were never populated, causing diagnostic data to be stuck at defaults
@@ -393,7 +295,10 @@ class ComputationEngine:
         # ---- Step DP: Inline DP optimizer (Phase 4, #441) ----
         # Runs before effective_cheap_price final update so solar_can_reach_target
         # is populated from DP result (not legacy solar-only simulation).
-        self._run_dp_optimizer_inline(data, now_dt)
+        config_options = self._build_optimizer_config_options()
+        self._optimizer_facade.run_inline(
+            data=data, now_dt=now_dt, config_options=config_options
+        )
 
         # ---- Step 6: boost_charge_needed (Phase 4: derive from DP decision) ----
         # Read from current-slot DP decision (not forecast_computer).
@@ -408,7 +313,12 @@ class ComputationEngine:
 
         # ---- Step 7: effective_cheap_price (final update) ----
         # Update effective_cheap_price with actual solar_can_reach_target from forecast
-        self._compute_effective_cheap_price(data, now_dt, before_dw, target_hour)
+        self._price_signals.compute_effective_cheap_price(
+            data=data,
+            now_dt=now_dt,
+            before_dw=before_dw,
+            target_hour=target_hour,
+        )
 
         # ---- Step 8: cheap_charge_stop_price ----
         # Hardcoded deadband (Issue #214)
@@ -417,39 +327,48 @@ class ComputationEngine:
 
         # ---- Step 4: solar_battery_forecast (legacy - for backwards compatibility) ----
         # Kept for API compatibility, but values derived from detailed forecast
-        self._compute_solar_battery_forecast(
-            data, now_dt, target_hour, before_dw, after_dw
+        self._forecast_pipeline.compute_solar_battery_forecast(
+            data=data,
+            now_dt=now_dt,
+            target_hour=target_hour,
+            before_dw=before_dw,
+            after_dw=after_dw,
+            target_pct=target_pct,
         )
 
         # ---- Step 9: forecast_spike_within_window ----
         # Hardcoded lookahead (Issue #214)
         lookahead = DEFAULT_FORECAST_LOOKAHEAD_HOURS
         cutoff = now_dt + timedelta(hours=lookahead)
-        data.forecast_spike_within_window = self._scan_forecast_for_spike(
+        data.forecast_spike_within_window = self._price_signals.scan_forecast_for_spike(
             data.feed_in_forecast, now_dt, cutoff
         )
         # max_forecast_price tracks the max SELL price (feed-in) for spike detection.
-        data.max_forecast_price = self._max_forecast_price(
+        data.max_forecast_price = self._price_signals.max_forecast_price(
             data.feed_in_forecast, now_dt, cutoff
         )
         # max_buy_forecast_price tracks the max BUY price for pre-charge decisions. (Fix #3)
-        data.max_buy_forecast_price = self._max_forecast_price(
+        data.max_buy_forecast_price = self._price_signals.max_forecast_price(
             data.general_forecast, now_dt, cutoff
         )
 
         # ---- Step 10: forecast_expensive_period_coming ----
-        data.forecast_expensive_period_coming = self._scan_forecast_for_spike(
-            data.general_forecast, now_dt, cutoff
+        data.forecast_expensive_period_coming = (
+            self._price_signals.scan_forecast_for_spike(
+                data.general_forecast, now_dt, cutoff
+            )
         )
 
         # ---- Step 10b: spike analysis (conservative mode) ----
-        self._analyze_spike(data, now_dt)
+        self._price_signals.analyze_spike(data, now_dt)
 
         # ---- Step 11: solar_weighted_avg_fit ----
-        self._compute_solar_weighted_avg_fit(data, now_dt, target_hour, after_dw)
+        self._forecast_pipeline.compute_solar_weighted_avg_fit(
+            data=data, now_dt=now_dt, target_hour=target_hour, after_dw=after_dw
+        )
 
         # ---- Step 12: decision_log ----
-        # Phase 4 (#441): active_mode is set by _run_dp_optimizer_inline (Phase 3).
+        # Phase 4 (#441): active_mode is set by the optimizer facade (Phase 3).
         # _compute_active_mode is removed in Phase 4.
 
         # Add entry when mode changes OR periodically for status updates
@@ -477,7 +396,7 @@ class ComputationEngine:
         # (computed earlier; left intentionally blank)
 
         # ---- Step 17: excess_solar_signals (backlog-high-017) ----
-        self._compute_excess_solar_signals(data, now_dt)
+        self._forecast_pipeline.compute_excess_solar_signals(data, now_dt)
 
         # ---- Step 18: weather correlation diagnostics (Issue #61) ----
         self._populate_weather_diagnostics(data)
@@ -498,244 +417,14 @@ class ComputationEngine:
         target_pct = float(
             self.entry.options.get(CONF_BATTERY_TARGET, DEFAULT_BATTERY_TARGET)
         )
-
-        if after_dw:
-            # After DW start: report current SOC, safe defaults
-            # Use DP decision if available (Phase 4, #441)
-            dw_entry = self._get_dp_decision_at_demand_window(data, target_hour)
-            if dw_entry:
-                predicted_soc = dw_entry["predicted_soc_pct"]
-                can_reach = predicted_soc >= target_pct
-            else:
-                # Fallback to current SOC (conservative - assumes no solar)
-                predicted_soc = data.soc
-                can_reach = data.soc >= target_pct
-
-            # Boost_needed indicates if solar alone can reach target (for dashboard display)
-            # After DW, boost is not applicable
-            boost_needed = False
-
-            # Mark target reached if SOC is there
-            target_reached = data.soc >= target_pct
-            if target_reached:
-                data.target_reached_today = True
-
-            # Calculate hours until next demand window (with day rollover for after_dw case)
-            next_dw_dt = now_dt.replace(
-                hour=target_hour, minute=0, second=0, microsecond=0
-            )
-            if next_dw_dt <= now_dt:
-                next_dw_dt += timedelta(days=1)
-            hours_to_next_dw = (next_dw_dt - now_dt).total_seconds() / 3600
-
-            data.solar_battery_forecast = {
-                "predicted_soc": round(predicted_soc, 1),
-                "solar_before_dw_kwh": 0.0,
-                "consumption_estimate_kwh": 0.0,
-                "net_solar_kwh": 0.0,
-                "deficit_kwh": 0.0,
-                "can_reach_target": can_reach,
-                "boost_needed": boost_needed,
-                "hours_to_target_time": round(hours_to_next_dw, 1),
-                "target_reached_today": target_reached,
-            }
-        else:
-            # Before DW: use DP decision (Phase 4, #441)
-            dw_entry = self._get_dp_decision_at_demand_window(data, target_hour)
-
-            if dw_entry:
-                # Use DP decision - includes grid charging effects
-                predicted_soc = dw_entry["predicted_soc_pct"]
-                can_reach = predicted_soc >= target_pct
-
-                # For boost_needed: calculate if solar ALONE can reach target
-                # (without grid charging) - this is the "solar gap" indicator
-                deficit_kwh = max(
-                    (target_pct - data.soc) / 100 * BATTERY_CAPACITY_KWH, 0
-                )
-
-                # Solar forecast: pessimistic estimate between now and DW
-                # Include tomorrow's forecast when target is tomorrow morning
-                all_solcast = [*data.solcast_today, *data.solcast_tomorrow]
-                solar_kwh = self._sum_solar_before_target(
-                    all_solcast, now_dt, target_hour
-                )
-
-                # Hours remaining until DW start
-                target_dt = now_dt.replace(
-                    hour=target_hour, minute=0, second=0, microsecond=0
-                )
-                hours_to_target = max((target_dt - now_dt).total_seconds() / 3600, 0)
-
-                # Consumption estimate from load_forecast_slots (Phase 4)
-                expected_load_kw = self._get_expected_load_kw_from_slots(
-                    data, hours_to_target
-                )
-                consumption_kwh = expected_load_kw * hours_to_target
-
-                # Net solar (after consumption) - solar only, no grid charging
-                net_solar = solar_kwh - consumption_kwh
-
-                # Boost needed if solar alone can't reach target.
-                # When allow_dw_entry_under_target is enabled, use solar_can_reach_target_in_dw
-                # which simulates solar charging through the DW period (not just to DW START).
-                # This ensures the diagnostic flag matches the actual decision logic.
-                allow_dw_under_target = self._get_switch_state(
-                    SWITCH_ALLOW_DW_ENTRY_UNDER_TARGET
-                )
-                if allow_dw_under_target and hasattr(
-                    data, "solar_can_reach_target_in_dw"
-                ):
-                    # Use the DW-extended simulation result for consistency
-                    boost_needed = (
-                        data.soc < target_pct and not data.solar_can_reach_target_in_dw
-                    )
-                else:
-                    # Standard calculation: boost needed if solar alone can't reach target before DW
-                    boost_needed = data.soc < target_pct and net_solar < deficit_kwh
-            else:
-                # Fallback if DP decision unavailable
-                # Hours remaining until DW start
-                target_dt = now_dt.replace(
-                    hour=target_hour, minute=0, second=0, microsecond=0
-                )
-                hours_to_target = max((target_dt - now_dt).total_seconds() / 3600, 0)
-
-                # Deficit: kWh needed to reach target
-                deficit_kwh = max(
-                    (target_pct - data.soc) / 100 * BATTERY_CAPACITY_KWH, 0
-                )
-
-                # Solar forecast: pessimistic estimate between now and DW
-                # Include tomorrow's forecast when target is tomorrow morning
-                all_solcast = [*data.solcast_today, *data.solcast_tomorrow]
-                solar_kwh = self._sum_solar_before_target(
-                    all_solcast, now_dt, target_hour
-                )
-
-                # Consumption estimate: current load extrapolated
-                expected_load_kw = self._get_expected_load_kw_from_slots(
-                    data, hours_to_target
-                )
-                consumption_kwh = expected_load_kw * hours_to_target
-
-                # Net solar (after consumption)
-                net_solar = solar_kwh - consumption_kwh
-
-                # Predicted SOC at DW (clamped to 0-100%)
-                net_solar_pct = net_solar / BATTERY_CAPACITY_KWH * 100
-                predicted_soc = max(0.0, min(100.0, data.soc + net_solar_pct))
-
-                # Can solar alone reach target? (fallback calculation)
-                can_reach = data.soc >= target_pct or net_solar >= deficit_kwh
-
-                # Boost needed? Only if gentle charging can't reach target before DW
-                if data.soc >= target_pct:
-                    boost_needed = False
-                else:
-                    remaining_deficit = max(deficit_kwh - max(net_solar, 0), 0)
-                    time_needed_hours = (
-                        remaining_deficit / (CHARGE_RATE_GRID_KW * 0.9)
-                        if remaining_deficit > 0
-                        else 0
-                    )
-                    boost_needed = (
-                        time_needed_hours > (hours_to_target - 0.5)
-                        and remaining_deficit > 0
-                    )
-
-            # Mark target reached if SOC is there
-            target_reached = data.soc >= target_pct
-            if target_reached:
-                data.target_reached_today = True
-
-            data.solar_battery_forecast = {
-                "predicted_soc": round(predicted_soc, 1),
-                "solar_before_dw_kwh": round(solar_kwh, 2),
-                "consumption_estimate_kwh": round(consumption_kwh, 2),
-                "net_solar_kwh": round(net_solar, 2),
-                "deficit_kwh": round(deficit_kwh, 2),
-                "can_reach_target": can_reach,
-                "boost_needed": boost_needed,
-                "hours_to_target_time": round(hours_to_target, 1),
-                "target_reached_today": target_reached,
-            }
-
-            # Store forecast history when hour changes (for planned vs actual chart)
-            current_hour = now_dt.hour
-            if (
-                self._last_forecast_hour is None
-                or current_hour != self._last_forecast_hour
-            ):
-                self._store_forecast_history(data, now_dt)
-                self._last_forecast_hour = current_hour
-
-    def _store_forecast_history(
-        self,
-        data: CoordinatorData,
-        now_dt: datetime,
-    ) -> None:
-        """Store DP optimizer predictions to history for planned vs actual comparison (Phase 4, #441).
-
-        Stores predictions at specific future times for accuracy tracking:
-        - What SOC we predict for 15 minutes from now
-        - What SOC we predict for 1 hour from now
-        - What SOC we predict for 4 hours from now
-
-        Each prediction has a target_time so we can later compare:
-        "What did we predict for time T?" vs "What was actual at time T?"
-        """
-        # Find DP decision entries for specific future times
-        slots = data.optimizer_decisions or []
-        if not slots:
-            return
-
-        # Store predictions for 15min, 1h, and 4h into the future
-        # These will be compared when that time arrives
-        for offset_minutes in [15, 60, 240]:
-            target_dt = now_dt + timedelta(minutes=offset_minutes)
-
-            # Normalize target_dt timezone for comparison
-            if target_dt.tzinfo is None:
-                target_dt = dt_util.as_local(dt_util.as_utc(target_dt))
-            else:
-                target_dt = dt_util.as_local(target_dt)
-
-            # Find the DP decision slot that covers target_dt
-            for slot in slots:
-                ts = slot.get("timestamp_iso", "")
-                if not ts:
-                    continue
-                try:
-                    slot_dt = datetime.fromisoformat(ts)
-                except ValueError:
-                    continue
-
-                # Normalize timezone
-                if slot_dt.tzinfo is None:
-                    slot_dt = dt_util.as_local(dt_util.as_utc(slot_dt))
-                else:
-                    slot_dt = dt_util.as_local(slot_dt)
-
-                # Check if this slot covers target_dt (within slot interval)
-                slot_interval = slot.get("slot_interval_minutes", 15)
-                slot_end = slot_dt + timedelta(minutes=slot_interval)
-
-                if slot_dt <= target_dt < slot_end:
-                    entry = {
-                        "prediction_time": now_dt.isoformat(),
-                        "target_time": target_dt.isoformat(),
-                        "offset_minutes": offset_minutes,
-                        "predicted_soc": slot.get("predicted_soc_pct", 0),
-                        "predicted_buy_price": slot.get("buy_price", 0),
-                        "predicted_sell_price": slot.get("sell_price", 0),
-                    }
-                    data.forecast_history.append(entry)
-                    break
-
-        # Keep last 200 entries (allows 4+ hours of predictions across multiple days)
-        if len(data.forecast_history) > 200:
-            data.forecast_history = data.forecast_history[-200:]
+        self._forecast_pipeline.compute_solar_battery_forecast(
+            data=data,
+            now_dt=now_dt,
+            target_hour=target_hour,
+            before_dw=before_dw,
+            after_dw=after_dw,
+            target_pct=target_pct,
+        )
 
     def _compute_load_forecast_slots(
         self,
@@ -744,39 +433,13 @@ class ComputationEngine:
         historical_avg_kw: dict[int, float],
         recent_load_kw: float,
     ) -> None:
-        """Populate data.load_forecast_slots with per-slot kW estimates.
-
-        Shared intermediate for DP optimizer slot building (extracted in #441 Phase 1).
-        Slots are fixed 15-min (TOTAL_SLOTS = 96) aligned to the current 5-min boundary.
-        """
-        current_5min = (now_dt.minute // 5) * 5
-        base_slot = now_dt.replace(minute=current_5min, second=0, microsecond=0)
-        current_hour = base_slot.hour
-
-        slots: list[float] = []
-        for i in range(TOTAL_SLOTS):
-            slot_start = base_slot + timedelta(minutes=15 * i)
-            slot_hour = slot_start.hour
-            hours_ahead = i / 4.0
-            temperature = data.weather_temperature_forecast.get(slot_hour)
-            load_kw, _ = self._load_forecaster.estimate_hourly_consumption_kw(
-                hourly_avg_kw=historical_avg_kw,
-                slot_hour=slot_hour,
-                current_hour=current_hour,
-                current_load_kw=data.load_power_kw,
-                recent_load_kw=recent_load_kw,
-                temperature=temperature,
-                hours_ahead=hours_ahead,
-            )
-            slots.append(load_kw)
-
-        data.load_forecast_slots = slots
-        _LOGGER.info(
-            "ISSUE_500 load_forecast_slots: %d slots, indices 4-8 = %s, recent_load=%.3f, hourly_avg_12=%.3f",
-            len(slots),
-            [round(slots[i], 3) for i in range(4, min(9, len(slots)))],
-            recent_load_kw,
-            historical_avg_kw.get(12, -1),
+        """Populate data.load_forecast_slots with per-slot kW estimates."""
+        self._forecast_pipeline.compute_load_forecast_slots(
+            data=data,
+            now_dt=now_dt,
+            historical_avg_kw=historical_avg_kw,
+            recent_load_kw=recent_load_kw,
+            total_slots=TOTAL_SLOTS,
         )
 
     def _compute_effective_cheap_price_preliminary(
@@ -788,7 +451,7 @@ class ComputationEngine:
         target_pct: float,
     ) -> None:
         """Compute preliminary effective cheap price threshold."""
-        self._price_calculator.compute_effective_cheap_price_preliminary(
+        self._price_signals.compute_effective_cheap_price_preliminary(
             data=data,
             now_dt=now_dt,
             before_dw=before_dw,
@@ -800,7 +463,7 @@ class ComputationEngine:
         self, data: CoordinatorData, now_dt: datetime, before_dw: bool, target_hour: int
     ) -> None:
         """Compute final effective cheap price threshold."""
-        self._price_calculator.compute_effective_cheap_price(
+        self._price_signals.compute_effective_cheap_price(
             data=data,
             now_dt=now_dt,
             before_dw=before_dw,
@@ -811,7 +474,7 @@ class ComputationEngine:
         self, data: CoordinatorData, now_dt: datetime, target_hour: int, after_dw: bool
     ) -> None:
         """Compute solar-weighted average feed-in tariff."""
-        self._price_calculator.compute_solar_weighted_avg_fit(
+        self._forecast_pipeline.compute_solar_weighted_avg_fit(
             data=data,
             now_dt=now_dt,
             target_hour=target_hour,
@@ -821,51 +484,10 @@ class ComputationEngine:
     def _get_dp_decision_at_demand_window(
         self, data: CoordinatorData, target_hour: int
     ) -> dict | None:
-        """Get the DP decision at or just after the demand window start time (Phase 4, #441).
-
-        Finds the first decision whose timestamp_iso is at or after the DW start
-        (target_hour:00:00). Handles post-DW period by looking for tomorrow's DW
-        if today's has passed.
-        """
-        from datetime import datetime, timedelta  # noqa: PLC0415
-
-        decisions = data.optimizer_decisions or []
-        if not decisions:
-            return None
-
-        # Normalise now to tz-aware local time
-        now_raw = dt_util.now()
-        now_local = (
-            dt_util.as_local(dt_util.as_utc(now_raw))
-            if now_raw.tzinfo is None
-            else dt_util.as_local(now_raw)
+        """Get the DP decision at or just after the demand window start time."""
+        return self._forecast_pipeline._get_dp_decision_at_demand_window(
+            data, target_hour
         )
-
-        # Calculate DW start datetime
-        dw_start_dt = now_local.replace(
-            hour=target_hour, minute=0, second=0, microsecond=0
-        )
-        if dw_start_dt <= now_local:
-            dw_start_dt += timedelta(days=1)
-
-        # Find first decision at or after DW start
-        for decision in decisions:
-            ts = decision.get("timestamp_iso", "")
-            if not ts:
-                continue
-            try:
-                slot_dt = datetime.fromisoformat(ts)
-            except ValueError:
-                continue
-            if slot_dt.tzinfo is None:
-                slot_local = dt_util.as_local(dt_util.as_utc(slot_dt))
-            else:
-                slot_local = dt_util.as_local(slot_dt)
-
-            if slot_local >= dw_start_dt:
-                return decision
-
-        return None
 
     def _get_expected_load_kw_from_slots(
         self, data: CoordinatorData, hours_to_target: float
@@ -875,13 +497,9 @@ class ComputationEngine:
         Averages slots from current slot to DW entry slot. Falls back to current
         load if no forecast slots available.
         """
-        if not data.load_forecast_slots:
-            return data.load_power_kw if data.load_power_kw > 0 else 0.5
-
-        # Number of 15-min slots until DW (4 slots per hour)
-        slots_until_dw = max(1, int(hours_to_target * 4))
-        relevant = data.load_forecast_slots[:slots_until_dw]
-        return sum(relevant) / len(relevant) if relevant else 0.5
+        return self._price_signals.get_expected_load_kw_from_slots(
+            data, hours_to_target
+        )
 
     # ========================================================================
     # PHASE 3 (#441): INLINE DP OPTIMIZER - ELIMINATE ONE-CYCLE LAG
@@ -916,204 +534,6 @@ class ComputationEngine:
             # via config_overrides), use it to avoid relying on dt_util.DEFAULT_TIME_ZONE.
             "ha_timezone": self.entry.options.get("ha_timezone", None),
         }
-
-    def _run_dp_optimizer_inline(self, data: CoordinatorData, now_dt: datetime) -> None:
-        """Run DP optimizer inline so active_mode has no cycle lag (Phase 3, #441).
-
-        Steps A–E from the architecture doc:
-          A: build slot contexts from raw data
-          B: run DPPlanner.plan()
-          C: write optimizer fields to data
-          D: derive apply plan from current-slot decision
-          E: assign data.active_mode (or default to SELF_CONSUMPTION on gate failure)
-        """
-        import uuid  # noqa: PLC0415
-
-        from homeassistant.util import dt as dt_util  # noqa: PLC0415
-
-        # Read config once for this cycle
-        config_options = self._build_optimizer_config_options()
-
-        try:
-            # Step A: Build slots from raw data (Phase 2 SlotBuilder)
-            # Allow config_options to override timezone (used in tests to inject the
-            # scenario's local timezone without needing to mock dt_util.DEFAULT_TIME_ZONE).
-            ha_timezone = config_options.get("ha_timezone") or (
-                str(dt_util.DEFAULT_TIME_ZONE)
-                if dt_util.DEFAULT_TIME_ZONE
-                else "Australia/Sydney"
-            )
-            slot_builder = SlotBuilder(
-                config_options=config_options, ha_timezone=ha_timezone
-            )
-            slots, slot_metadata = slot_builder.build_slots(
-                data, data.adaptive_params, now_dt=now_dt
-            )
-
-            # Record forecasts for solar accuracy tracking (Issue #513)
-            weather_condition = getattr(data, "weather_condition", None) or "unknown"
-            self._record_forecasts_for_slots(slots, weather_condition)
-
-            # Apply bias correction to solar forecasts if tracker has data (Issue #513)
-            self._apply_bias_correction_to_slots(slots, weather_condition)
-
-            if not slots:
-                _LOGGER.warning("DP optimizer: no slots available, skipping")
-                return
-
-            # Step B: Build OptimizerConfig
-            optimizer_config = _build_optimizer_config(data, config_options)
-
-            # Normalize and validate SOC
-            initial_soc, soc_info = _normalize_initial_soc(data.soc, optimizer_config)
-            if initial_soc is None:
-                _LOGGER.warning(
-                    "DP optimizer: invalid SOC %s, skipping", soc_info.get("error")
-                )
-                return
-
-            # Step B: Run DPPlanner
-            cycle_id = uuid.uuid4().hex[:12]
-            from .computation_engine_lib.optimizer_dp import (
-                OptimizerInputs,  # noqa: PLC0415
-            )
-
-            inputs = OptimizerInputs(
-                cycle_id=cycle_id,
-                initial_soc_pct=initial_soc,
-                slots=slots,
-                config=optimizer_config,
-            )
-            result = self._dp_planner.plan(inputs)
-
-            # Step C: Write optimizer fields (always, even on solve failure, for diagnostics)
-            self._write_optimizer_fields(
-                data, result, slot_metadata, config_options, cycle_id
-            )
-
-            # Steps D+E: Derive apply plan and assign active_mode
-            self._assign_active_mode(data, result, optimizer_config)
-
-        except Exception as e:
-            _LOGGER.warning(
-                "Inline DP optimizer failed (non-blocking): %s", e, exc_info=True
-            )
-            # active_mode falls through to legacy _compute_active_mode
-
-    def _write_optimizer_fields(
-        self,
-        data: CoordinatorData,
-        result: Any,  # OptimizerResult
-        slot_metadata: Any,  # SlotBuildMetadata
-        config_options: dict[str, Any],
-        cycle_id: str,
-    ) -> None:
-        """Write DP result to CoordinatorData fields (Step C)."""
-        from homeassistant.util import dt as dt_util  # noqa: PLC0415
-
-        data.optimizer_result = _serialize_result(result)
-        data.optimizer_decisions = [_serialize_decision(d) for d in result.decisions]
-        data.optimizer_summary = _build_summary(
-            result=result,
-            cycle_id=cycle_id,
-            cycle_timestamp_iso=dt_util.utcnow().isoformat(),
-            parity_info=slot_metadata.to_parity_dict(),
-            config_options=config_options,
-        )
-
-        # Update forecast horizon from slot metadata (Issue #522)
-        # This bridges the actual horizon computed by slot_builder to the optimizer config
-        data.forecast_horizon_hours = slot_metadata.horizon_hours
-
-        # Derive solar_can_reach_target from DP result (Phase 4, #441)
-        # Replaces the legacy _simulate_future_soc_with_solar_only() calls.
-        data.solar_can_reach_target = result.can_solar_reach_target
-        # solar_can_reach_target_in_dw: when allow_dw_entry_under_target is True,
-        # terminal_penalty_idx is DW end — so can_solar_reach_target already reflects
-        # the extended horizon. Mirror the value and set False when the switch is off
-        # (same as the legacy path which set it to False when allow_dw_under_target=False).
-        allow_dw_under_target = config_options.get("allow_dw_entry_under_target", False)
-        data.solar_can_reach_target_in_dw = (
-            result.can_solar_reach_target if allow_dw_under_target else False
-        )
-
-    def _assign_active_mode(
-        self,
-        data: CoordinatorData,
-        result: Any,  # OptimizerResult
-        optimizer_config: Any,  # OptimizerConfig
-    ) -> None:
-        """Assign data.active_mode from DP result within the same cycle (Steps D+E).
-
-        On safety gate failure, defaults to SELF_CONSUMPTION — the safe hardware state.
-        """
-        # Build alignment info from the summary we just wrote
-        alignment = {
-            "valid": True,  # SlotBuilder produces aligned slots by construction (Phase 2)
-            "issues": [],
-            "warnings": [],
-        }
-
-        config_options = self._build_optimizer_config_options()
-        safety_gate = OptimizerSafetyGate(config_options)
-        gate_result = safety_gate.check_admission(data, result, alignment)
-
-        if not gate_result.allowed:
-            _LOGGER.info(
-                "DP optimizer safety gate blocked: %s — defaulting to SELF_CONSUMPTION",
-                gate_result.block_reason,
-            )
-
-            data.active_mode = _BatteryMode.SELF_CONSUMPTION
-            data.optimizer_last_apply_status = "blocked"
-            data.optimizer_safety_block_reason = gate_result.block_reason or ""
-            _LOGGER.warning(
-                "Optimizer safety gate failed — defaulting to SELF_CONSUMPTION"
-            )
-            return
-
-        # Gate passed — derive apply plan from current slot
-        current_slot_idx = _find_current_slot_index(data)
-        apply_plan = _derive_runtime_apply_plan(
-            data.optimizer_decisions, current_slot_idx, optimizer_config
-        )
-        data.optimizer_apply_plan = apply_plan
-
-        # Map battery_mode string → BatteryMode enum
-        battery_mode_str = apply_plan.get("battery_mode", "")
-        try:
-            new_mode = _BatteryMode(battery_mode_str)
-            # Record decision timestamp when mode changes (Issue #501)
-            if new_mode != data.active_mode:
-                from homeassistant.util import dt as dt_util  # noqa: PLC0415
-
-                decision_time = dt_util.now()
-                if decision_time is not None:
-                    data.decision_timestamp = decision_time
-                    data.decision_mode = new_mode
-                    _LOGGER.info(
-                        "Decision lag tracking: mode change %s → %s at %s",
-                        data.active_mode.value,
-                        new_mode.value,
-                        decision_time.isoformat(),
-                    )
-            data.active_mode = new_mode
-            data.optimizer_last_apply_status = "ready_to_apply"
-            data.optimizer_safety_block_reason = ""
-            _LOGGER.info(
-                "DP optimizer: selected %s (action=%s, slot=%d)",
-                battery_mode_str,
-                apply_plan.get("action"),
-                current_slot_idx,
-            )
-        except ValueError:
-            _LOGGER.warning(
-                "DP optimizer: invalid battery_mode '%s' — defaulting to SELF_CONSUMPTION",
-                battery_mode_str,
-            )
-
-            data.active_mode = _BatteryMode.SELF_CONSUMPTION
-            data.optimizer_last_apply_status = "fallback"
 
     def _add_to_decision_log(
         self, data: CoordinatorData, now_dt: datetime, mode_change: bool
@@ -1284,139 +704,21 @@ class ComputationEngine:
 
     async def async_initialize_forecast_history_storage(self) -> None:
         """Initialize forecast history storage and load persisted data."""
-        try:
-            from homeassistant.helpers.storage import Store
-
-            self._forecast_history_store = Store(
-                self.hass, 1, "localshift_forecast_history"
-            )
-            _LOGGER.info("Forecast history storage initialized")
-        except Exception as e:
-            _LOGGER.warning("Failed to initialize forecast history storage: %s", e)
-            self._forecast_history_store = None
+        await self._forecast_history_store.async_initialize()
 
     async def async_load_forecast_history(self, data: CoordinatorData) -> None:
         """Load persisted forecast history from storage.
 
         Called during coordinator startup to restore predictions across restarts.
         """
-        if self._forecast_history_store is None:
-            _LOGGER.debug("No forecast history store available")
-            return
-
-        try:
-            stored_data = await self._forecast_history_store.async_load()
-            if not stored_data or not isinstance(stored_data, dict):
-                return
-
-            history = stored_data.get("forecast_history", [])
-            first_prediction = stored_data.get("first_prediction_time", "")
-
-            if history:
-                now_dt = dt_util.now()
-                cutoff = now_dt - timedelta(hours=4)
-                valid_entries = self._filter_valid_history_entries(history, cutoff)
-
-                data.forecast_history = valid_entries
-                data.forecast_first_prediction_time = first_prediction
-                data.forecast_history_count = len(valid_entries)
-
-                _LOGGER.info(
-                    "Loaded %d forecast history entries from storage (filtered from %d)",
-                    len(valid_entries),
-                    len(history),
-                )
-
-                if not first_prediction and valid_entries:
-                    self._find_first_prediction_time(data, valid_entries)
-
-            self._forecast_history_loaded = True
-        except Exception as e:
-            _LOGGER.warning("Failed to load forecast history: %s", e)
-
-    def _filter_valid_history_entries(
-        self, history: list[dict], cutoff: datetime
-    ) -> list[dict]:
-        """Filter history entries to only valid ones within cutoff time.
-
-        Args:
-            history: List of history entry dictionaries
-            cutoff: Cutoff datetime for filtering
-
-        Returns:
-            List of valid entries
-        """
-        valid_entries = []
-        for entry in history:
-            if "target_time" not in entry:
-                continue
-            if self._is_history_entry_valid(entry, cutoff):
-                valid_entries.append(entry)
-        return valid_entries
-
-    def _is_history_entry_valid(self, entry: dict, cutoff: datetime) -> bool:
-        """Check if a history entry is valid and within cutoff time.
-
-        Args:
-            entry: History entry dictionary
-            cutoff: Cutoff datetime
-
-        Returns:
-            True if entry is valid and within cutoff
-        """
-        try:
-            target_dt = datetime.fromisoformat(entry["target_time"])
-            if target_dt.tzinfo is None:
-                target_dt = dt_util.as_local(dt_util.as_utc(target_dt))
-            else:
-                target_dt = dt_util.as_local(target_dt)
-            return target_dt >= cutoff
-        except (ValueError, TypeError):
-            return False
-
-    def _find_first_prediction_time(
-        self, data: CoordinatorData, entries: list[dict]
-    ) -> None:
-        """Find first prediction time from loaded entries.
-
-        Args:
-            data: CoordinatorData to update
-            entries: List of valid history entries
-        """
-        for entry in entries:
-            if entry.get("prediction_time"):
-                data.forecast_first_prediction_time = entry["prediction_time"]
-                break
+        await self._forecast_history_store.async_load(data)
 
     async def async_save_forecast_history(self, data: CoordinatorData) -> None:
         """Persist forecast history to storage.
 
         Called after new predictions are stored.
         """
-        if self._forecast_history_store is None:
-            return
-
-        try:
-            # Store only entries with target_time (not legacy entries)
-            entries_to_save = [
-                entry
-                for entry in data.forecast_history
-                if "target_time" in entry and "offset_minutes" in entry
-            ]
-
-            # Limit to recent entries (keep storage size manageable)
-            if len(entries_to_save) > 100:
-                entries_to_save = entries_to_save[-100:]
-
-            stored_data = {
-                "forecast_history": entries_to_save,
-                "first_prediction_time": data.forecast_first_prediction_time,
-            }
-
-            await self._forecast_history_store.async_save(stored_data)
-            _LOGGER.debug("Saved %d forecast history entries", len(entries_to_save))
-        except Exception as e:
-            _LOGGER.warning("Failed to save forecast history: %s", e)
+        await self._forecast_history_store.async_save(data)
 
     # ========================================================================
     # WEATHER CORRELATION (Issue #61)
@@ -1537,7 +839,7 @@ class ComputationEngine:
         now_dt: datetime,
     ) -> None:
         """Analyze feed-in forecast for spike window details."""
-        self._spike_analyzer.analyze_spike(data, now_dt)
+        self._price_signals.analyze_spike(data, now_dt)
 
     # ========================================================================
     # EXCESS SOLAR LOAD SHIFTING (backlog-high-017)
@@ -1549,7 +851,7 @@ class ComputationEngine:
         now_dt: datetime,
     ) -> None:
         """Compute excess solar load shifting signals."""
-        self._excess_solar_signals.compute_signals(data, now_dt)
+        self._forecast_pipeline.compute_excess_solar_signals(data, now_dt)
 
     # ========================================================================
     # FORECAST ACCURACY TRACKING (Issue #37 Phase 2)
