+++ custom_components/localshift/computation_engine_lib/optimizer_facade.py
@@ -1,0 +1,252 @@
+"""Optimizer facade for DP planner/runner orchestration."""
+
+from __future__ import annotations
+
+import logging
+import uuid
+from datetime import datetime
+from typing import Any
+
+from homeassistant.util import dt as dt_util
+
+from ..const import BatteryMode as _BatteryMode
+from ..coordinator_data import CoordinatorData
+from .optimizer_dp import DPPlanner, OptimizerInputs
+from .optimizer_runner import (
+    OptimizerSafetyGate,
+    _build_optimizer_config,
+    _build_summary,
+    _derive_runtime_apply_plan,
+    _find_current_slot_index,
+    _normalize_initial_soc,
+    _serialize_decision,
+    _serialize_result,
+)
+from .slot_builder import SlotBuilder
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class OptimizerFacade:
+    """Facade that runs the DP optimizer and writes results to CoordinatorData."""
+
+    def __init__(
+        self,
+        planner: DPPlanner | None = None,
+        slot_builder_cls: type[Any] = SlotBuilder,
+    ) -> None:
+        self._planner = planner or DPPlanner()
+        self._slot_builder_cls = slot_builder_cls
+        self._solar_accuracy_tracker: Any = None
+
+    def set_solar_accuracy_tracker(self, tracker: Any) -> None:
+        """Set the solar accuracy tracker for bias correction."""
+        self._solar_accuracy_tracker = tracker
+
+    def _record_forecasts_for_slots(
+        self, slots: list[Any], weather_condition: str
+    ) -> None:
+        if self._solar_accuracy_tracker is None:
+            return
+
+        recorded = 0
+        for slot in slots:
+            if slot.solar_kwh > 0.01:
+                period_start = datetime.fromisoformat(slot.timestamp_iso)
+                self._solar_accuracy_tracker.record_forecast(
+                    period_start=period_start,
+                    forecast_kwh=slot.solar_kwh,
+                    weather_condition=weather_condition,
+                )
+                recorded += 1
+
+        if recorded > 0:
+            _LOGGER.debug("Recorded %d solar forecasts for accuracy tracking", recorded)
+
+    def _apply_bias_correction_to_slots(
+        self, slots: list[Any], weather_condition: str
+    ) -> None:
+        if self._solar_accuracy_tracker is None:
+            return
+
+        time_of_day = "day" if 6 <= dt_util.now().hour < 18 else "night"
+        now = dt_util.now()
+        month = now.month
+        if month in (12, 1, 2):
+            season = "summer"
+        elif month in (3, 4, 5):
+            season = "autumn"
+        elif month in (6, 7, 8):
+            season = "winter"
+        else:
+            season = "spring"
+
+        bias_multiplier = self._solar_accuracy_tracker.get_bias_correction(
+            time_of_day, weather_condition, season
+        )
+
+        if bias_multiplier != 1.0:
+            corrected = 0
+            for slot in slots:
+                if slot.solar_kwh > 0.01:
+                    original = slot.solar_kwh
+                    slot.solar_kwh = max(0.0, slot.solar_kwh * bias_multiplier)
+                    if abs(slot.solar_kwh - original) > 0.001:
+                        corrected += 1
+            if corrected > 0:
+                _LOGGER.info(
+                    "Applied bias correction: multiplier=%.2f (%s/%s/%s), corrected %d slots",
+                    bias_multiplier,
+                    time_of_day,
+                    weather_condition,
+                    season,
+                    corrected,
+                )
+
+    def run_inline(
+        self, data: CoordinatorData, now_dt: Any, config_options: dict[str, Any]
+    ) -> None:
+        """Run DP optimizer inline so active_mode has no cycle lag."""
+        try:
+            ha_timezone = config_options.get("ha_timezone") or (
+                str(dt_util.DEFAULT_TIME_ZONE)
+                if dt_util.DEFAULT_TIME_ZONE
+                else "Australia/Sydney"
+            )
+            slot_builder = self._slot_builder_cls(
+                config_options=config_options, ha_timezone=ha_timezone
+            )
+            slots, slot_metadata = slot_builder.build_slots(
+                data, data.adaptive_params, now_dt=now_dt
+            )
+
+            weather_condition = getattr(data, "weather_condition", None) or "unknown"
+            self._record_forecasts_for_slots(slots, weather_condition)
+            self._apply_bias_correction_to_slots(slots, weather_condition)
+
+            if not slots:
+                _LOGGER.warning("DP optimizer: no slots available, skipping")
+                return
+
+            optimizer_config = _build_optimizer_config(data, config_options)
+
+            initial_soc, soc_info = _normalize_initial_soc(data.soc, optimizer_config)
+            if initial_soc is None:
+                _LOGGER.warning(
+                    "DP optimizer: invalid SOC %s, skipping", soc_info.get("error")
+                )
+                return
+
+            cycle_id = uuid.uuid4().hex[:12]
+            inputs = OptimizerInputs(
+                cycle_id=cycle_id,
+                initial_soc_pct=initial_soc,
+                slots=slots,
+                config=optimizer_config,
+            )
+            result = self._planner.plan(inputs)
+
+            self._write_optimizer_fields(
+                data, result, slot_metadata, config_options, cycle_id
+            )
+
+            self._assign_active_mode(data, result, optimizer_config, config_options)
+
+        except Exception as exc:  # noqa: BLE001
+            _LOGGER.warning(
+                "Inline DP optimizer failed (non-blocking): %s", exc, exc_info=True
+            )
+
+    def _write_optimizer_fields(
+        self,
+        data: CoordinatorData,
+        result: Any,
+        slot_metadata: Any,
+        config_options: dict[str, Any],
+        cycle_id: str,
+    ) -> None:
+        data.optimizer_result = _serialize_result(result)
+        data.optimizer_decisions = [_serialize_decision(d) for d in result.decisions]
+        data.optimizer_summary = _build_summary(
+            result=result,
+            cycle_id=cycle_id,
+            cycle_timestamp_iso=dt_util.utcnow().isoformat(),
+            parity_info=slot_metadata.to_parity_dict(),
+            config_options=config_options,
+        )
+
+        data.forecast_horizon_hours = slot_metadata.horizon_hours
+
+        data.solar_can_reach_target = result.can_solar_reach_target
+        allow_dw_under_target = config_options.get("allow_dw_entry_under_target", False)
+        data.solar_can_reach_target_in_dw = (
+            result.can_solar_reach_target if allow_dw_under_target else False
+        )
+
+    def _assign_active_mode(
+        self,
+        data: CoordinatorData,
+        result: Any,
+        optimizer_config: Any,
+        config_options: dict[str, Any],
+    ) -> None:
+        alignment = {
+            "valid": True,
+            "issues": [],
+            "warnings": [],
+        }
+
+        safety_gate = OptimizerSafetyGate(config_options)
+        gate_result = safety_gate.check_admission(data, result, alignment)
+
+        if not gate_result.allowed:
+            _LOGGER.info(
+                "DP optimizer safety gate blocked: %s — defaulting to SELF_CONSUMPTION",
+                gate_result.block_reason,
+            )
+
+            data.active_mode = _BatteryMode.SELF_CONSUMPTION
+            data.optimizer_last_apply_status = "blocked"
+            data.optimizer_safety_block_reason = gate_result.block_reason or ""
+            _LOGGER.warning(
+                "Optimizer safety gate failed — defaulting to SELF_CONSUMPTION"
+            )
+            return
+
+        current_slot_idx = _find_current_slot_index(data)
+        apply_plan = _derive_runtime_apply_plan(
+            data.optimizer_decisions, current_slot_idx, optimizer_config
+        )
+        data.optimizer_apply_plan = apply_plan
+
+        battery_mode_str = apply_plan.get("battery_mode", "")
+        try:
+            new_mode = _BatteryMode(battery_mode_str)
+            if new_mode != data.active_mode:
+                decision_time = dt_util.now()
+                if decision_time is not None:
+                    data.decision_timestamp = decision_time
+                    data.decision_mode = new_mode
+                    _LOGGER.info(
+                        "Decision lag tracking: mode change %s → %s at %s",
+                        data.active_mode.value,
+                        new_mode.value,
+                        decision_time.isoformat(),
+                    )
+            data.active_mode = new_mode
+            data.optimizer_last_apply_status = "ready_to_apply"
+            data.optimizer_safety_block_reason = ""
+            _LOGGER.info(
+                "DP optimizer: selected %s (action=%s, slot=%d)",
+                battery_mode_str,
+                apply_plan.get("action"),
+                current_slot_idx,
+            )
+        except ValueError:
+            _LOGGER.warning(
+                "DP optimizer: invalid battery_mode '%s' — defaulting to SELF_CONSUMPTION",
+                battery_mode_str,
+            )
+
+            data.active_mode = _BatteryMode.SELF_CONSUMPTION
+            data.optimizer_last_apply_status = "fallback"
