+++ tests/test_computation_engine.py
@@ -1,6 +1,6 @@
 """Unit tests for ComputationEngine."""
 
-from datetime import datetime, time, timedelta, timezone
+from datetime import datetime, time
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -265,148 +265,6 @@ def test_decision_log_periodic_update(computation_engine, coordinator_data):
     assert initial_log_length >= 1
     # First entry should be initial status
     assert "reason" in coordinator_data.decision_log[-1]
-
-
-# =============================================================================
-# FORECAST CHANGE TRACKER TESTS
-# =============================================================================
-
-
-class TestSpikeAnalysis:
-    """Tests for _analyze_spike conservative mode logic."""
-
-    def test_analyze_spike_disabled(self, computation_engine, coordinator_data):
-        """Spike analysis should skip when conservative mode disabled."""
-        computation_engine._get_switch_state = MagicMock(return_value=False)
-
-        now_dt = datetime(2026, 2, 16, 18, 0, 0, tzinfo=timezone(timedelta(hours=11)))
-        computation_engine._analyze_spike(coordinator_data, now_dt)
-
-        # Should not populate spike fields
-        assert coordinator_data.spike_in_conservative_mode is False
-        assert coordinator_data.spike_end_time is None
-
-    def test_analyze_spike_no_spike_in_forecast(
-        self, computation_engine, coordinator_data
-    ):
-        """Spike analysis should handle no spike in forecast."""
-
-        # Enable conservative mode
-        def mock_switch_state(key):
-            if key == "spike_discharge_conservative":
-                return True
-            return False
-
-        computation_engine._get_switch_state = MagicMock(side_effect=mock_switch_state)
-
-        # No spike prices in forecast
-        coordinator_data.feed_in_forecast = [
-            {
-                "start_time": "2026-02-16T18:00:00+11:00",
-                "end_time": "2026-02-16T18:05:00+11:00",
-                "per_kwh": 0.08,
-            },
-            {
-                "start_time": "2026-02-16T18:05:00+11:00",
-                "end_time": "2026-02-16T18:10:00+11:00",
-                "per_kwh": 0.09,
-            },
-        ]
-
-        now_dt = datetime(2026, 2, 16, 18, 0, 0, tzinfo=timezone(timedelta(hours=11)))
-        computation_engine._analyze_spike(coordinator_data, now_dt)
-
-        # Should not detect spike
-        assert coordinator_data.spike_end_time is None
-        assert coordinator_data.spike_max_price == 0.0
-
-    def test_analyze_spike_with_spike_prices(
-        self, computation_engine, coordinator_data
-    ):
-        """Spike analysis should detect and analyze spike prices."""
-
-        # Enable conservative mode
-        def mock_switch_state(key):
-            if key == "spike_discharge_conservative":
-                return True
-            return False
-
-        computation_engine._get_switch_state = MagicMock(side_effect=mock_switch_state)
-        # Also mock the spike analyzer's reference to the switch state function
-        computation_engine._spike_analyzer._get_switch_state = MagicMock(
-            side_effect=mock_switch_state
-        )
-
-        # Create forecast with spike prices (> 1.0 $/kWh)
-        # Note: spike_status field is required for spike detection
-        coordinator_data.feed_in_forecast = [
-            {
-                "start_time": "2026-02-16T18:00:00+11:00",
-                "end_time": "2026-02-16T18:05:00+11:00",
-                "per_kwh": 1.50,  # Spike price
-                "spike_status": "spike",  # Required for spike detection
-            },
-            {
-                "start_time": "2026-02-16T18:05:00+11:00",
-                "end_time": "2026-02-16T18:10:00+11:00",
-                "per_kwh": 2.00,  # Higher spike
-                "spike_status": "spike",  # Required for spike detection
-            },
-            {
-                "start_time": "2026-02-16T18:10:00+11:00",
-                "end_time": "2026-02-16T18:15:00+11:00",
-                "per_kwh": 0.10,  # Normal price (no spike_status)
-            },
-        ]
-
-        now_dt = datetime(2026, 2, 16, 18, 0, 0, tzinfo=timezone(timedelta(hours=11)))
-
-        # Mock the _analyze_spike_window utility function to return spike data
-        spike_end = datetime(
-            2026, 2, 16, 18, 10, 0, tzinfo=timezone(timedelta(hours=11))
-        )
-        max_price = 2.0
-        spike_prices = [1.50, 2.00]
-        computation_engine._spike_analyzer._analyze_spike_window = MagicMock(
-            return_value=(spike_end, max_price, spike_prices)
-        )
-
-        computation_engine._analyze_spike(coordinator_data, now_dt)
-
-        # Should detect spike
-        assert coordinator_data.spike_in_conservative_mode is True
-        assert coordinator_data.spike_max_price > 1.0
-        assert coordinator_data.spike_price_threshold > 0
-
-    def test_analyze_spike_calculates_reserve_soc(
-        self, computation_engine, coordinator_data
-    ):
-        """Spike analysis should calculate reserve SOC needed."""
-
-        # Enable conservative mode
-        def mock_switch_state(key):
-            if key == "spike_discharge_conservative":
-                return True
-            return False
-
-        computation_engine._get_switch_state = MagicMock(side_effect=mock_switch_state)
-
-        # Create forecast with spike
-        coordinator_data.feed_in_forecast = [
-            {
-                "start_time": "2026-02-16T18:00:00+11:00",
-                "end_time": "2026-02-16T18:30:00+11:00",
-                "per_kwh": 2.00,
-            },
-        ]
-        coordinator_data.load_power_kw = 1.0  # 1 kW load
-
-        now_dt = datetime(2026, 2, 16, 18, 0, 0, tzinfo=timezone(timedelta(hours=11)))
-        computation_engine._analyze_spike(coordinator_data, now_dt)
-
-        # Should calculate reserve SOC
-        assert coordinator_data.spike_reserve_soc >= 0
-        assert coordinator_data.spike_hours_remaining >= 0
 
 
 # =============================================================================
